% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-km.R
\docType{data}
\name{GeomIcens}
\alias{GeomIcens}
\title{Display Kaplan Meier Curve}
\format{
An object of class \code{GeomIcens} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomIcens
}
\description{
Kaplan Meier Curve
}
\keyword{datasets}
